feat!: decouple SQLite/Postgres entry points from native dependencies
The ./sqlite entry point had a top-level `import Database from
"better-sqlite3"` which poisoned it for non-Node bundlers (Cloudflare
Workers, D1, Durable Objects). Move createLocalSqliteBackend to
./sqlite/local so that ./sqlite and ./postgres are clean, portable
entry points exporting only Drizzle backends and DDL generators.

- Rename test-helpers.ts → ddl.ts (production code, not test infra)
- Rename getSqliteMigrationSQL → generateSqliteMigrationSQL
- Rename getPostgresMigrationSQL → generatePostgresMigrationSQL
- Create ./sqlite/local for better-sqlite3-dependent code
- Remove ./drizzle/* entry points (consolidated into ./sqlite, ./postgres)
- Remove unused individual table type aliases from schema modules

BREAKING CHANGE: ./drizzle/* entry points removed, createLocalSqliteBackend
moved to ./sqlite/local, migration SQL generators renamed.
