+++ apps/docs/src/content/docs/performance/indexes.md
@@ -22,7 +22,7 @@ Define your indexes once and pass them into the Drizzle schema factories:
 
 ```ts
 import { defineEdge, defineNode } from "@nicia-ai/typegraph";
-import { createPostgresTables } from "@nicia-ai/typegraph/drizzle/schema/postgres";
+import { createPostgresTables } from "@nicia-ai/typegraph/postgres";
 import { andWhere, defineEdgeIndex, defineNodeIndex } from "@nicia-ai/typegraph/indexes";
 import { z } from "zod";
 
@@ -62,7 +62,7 @@ export const typegraphTables = createPostgresTables({}, {
 For SQLite, use `createSqliteTables`:
 
 ```ts
-import { createSqliteTables } from "@nicia-ai/typegraph/drizzle/schema/sqlite";
+import { createSqliteTables } from "@nicia-ai/typegraph/sqlite";
 
 export const typegraphTables = createSqliteTables({}, {
   indexes: [personEmail, worksAtRoleOut],
