+++ package.json
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@changesets/changelog-github": "^0.5.2",
     "@changesets/cli": "^2.29.8",
-    "knip": "^5.84.1",
+    "knip": "^5.85.0",
     "markdownlint-cli2": "^0.21.0",
     "npm-run-all2": "^8.0.4",
     "prettier": "^3.8.1",
