+++ packages/typegraph/tsup.config.ts
@@ -8,12 +8,8 @@ export default defineConfig({
     "schema/index": "src/schema/index.ts",
     "indexes/index": "src/indexes/index.ts",
     "backend/sqlite/index": "src/backend/sqlite/index.ts",
+    "backend/sqlite/local": "src/backend/sqlite/local.ts",
     "backend/postgres/index": "src/backend/postgres/index.ts",
-    "backend/drizzle/index": "src/backend/drizzle/index.ts",
-    "backend/drizzle/sqlite": "src/backend/drizzle/sqlite.ts",
-    "backend/drizzle/postgres": "src/backend/drizzle/postgres.ts",
-    "backend/drizzle/schema/sqlite": "src/backend/drizzle/schema/sqlite.ts",
-    "backend/drizzle/schema/postgres": "src/backend/drizzle/schema/postgres.ts",
   },
   format: ["esm", "cjs"],
   dts: true,
